fix sys/socket sendrecv return types.

+++ src/sys/socket/recv.cpp
@@ -29,6 +29,6 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <sys/socket.h>
 #include "basic_impl.hpp"
-IFT_BASIC_IMPL(int, recv, (int, void*, size_t, int), -1,
+IFT_BASIC_IMPL(ssize_t, recv, (int, void*, size_t, int), -1,
                (int sockfd, void* buf, size_t len, int flags),
                (sockfd, buf, len, flags), )
