+++ src/sys/socket/send.cpp
@@ -29,6 +29,6 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <sys/socket.h>
 #include "basic_impl.hpp"
-IFT_BASIC_IMPL(int, send, (int, const void*, size_t, int), -1,
+IFT_BASIC_IMPL(ssize_t, send, (int, const void*, size_t, int), -1,
                (int sockfd, const void* buf, size_t len, int flags),
                (sockfd, buf, len, flags), )
