+++ src/unistd/readlinkat.cpp
@@ -29,6 +29,6 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <unistd.h>
 #include "basic_impl.hpp"
-IFT_BASIC_IMPL(int, readlinkat, (int, const char*, char*, size_t), -1,
+IFT_BASIC_IMPL(ssize_t, readlinkat, (int, const char*, char*, size_t), -1,
                (int dirfd, const char* path, char* buf, size_t bufsiz),
                (dirfd, path, buf, bufsiz), throw ())
